Answer the self-review on the read swap
Findings from the judgement pass, all verified against the code:

The status mapping had two rows wrong. A staged deletion (`1 D.`) read
as "present and different" for a file that is gone from disk, which
would have sent stageWorktree down the add branch on a missing file; and
intent-to-add (`.A`) read as "in HEAD". Git also reports a path removed
from the index but kept on disk twice, as `1 D.` and `?`; that is one row
now, and in the scan against a branch point the untracked list overrides
the diff's D for the same path, because a patch that carried that
deletion would delete a file the contributor still has (#85).

collectChangedFiles swallowed a failed resolveRef and went on to report
"No changes" for a site whose .git is gone; it throws instead. The
cat-file parser stops on a missing echo whose path holds a newline
rather than poisoning every later offset.

The docs no longer claim that every read runs on the binary: the reads
inside the write flows stay on isomorphic-git until #385. The Git tests
are split by layer as TESTING.md asks, statusRows and changesAgainst
take an injected runner so a test proves the Windows autocrlf view
reaches the status and diff argv, and TESTING.md names the shared helper.
