+++ tests/unit/git-read.test.cjs
@@ -1,17 +1,13 @@
 const test = require('node:test');
 const assert = require('node:assert/strict');
-const fs = require('node:fs');
-const path = require('node:path');
 
 const read = require('../../src/git-read.cjs');
-const { git, tempDir } = require('./helpers/git.cjs');
 
-// Two halves. The parsers are pure functions over bytes, so the first half
-// feeds them the exact byte layouts Git documents for each porcelain format,
-// including the shapes that only a hostile repository produces (a stray
-// rename entry, a NUL inside a blob, a path with a space). The second half
-// runs the read functions against a repository the bundled Git built, so the
-// flag set each command uses is proven on this platform too.
+// The parsers are pure functions over bytes, so these feed them the exact
+// byte layouts Git documents for each porcelain format, including the shapes
+// only a hostile repository produces (a stray rename entry, a NUL inside a
+// blob, a path with a space). The same commands against a repository the
+// bundled Git built are in git-read.integration.test.cjs.
 
 const z = (...fields) => Buffer.from(fields.map((f) => `${f}\0`).join(''), 'utf8');
 
@@ -26,9 +22,11 @@ test('status v2: every XY combination lands on the documented row', () => {
 		['A.', [0, 2, 2]],
 		['AM', [0, 2, 3]],
 		['AD', [0, 0, 3]],
-		['D.', [1, 2, 0]],
+		['D.', [1, 0, 0]],
 		['T.', [1, 2, 2]],
-		['.T', [1, 2, 1]]
+		['.T', [1, 2, 1]],
+		// Intent-to-add: not in HEAD, on disk, a placeholder in the index.
+		['.A', [0, 2, 0]]
 	];
 	for (const [xy, expected] of cases) {
 		const [, head, workdir, stage] = read.rowFromStatusEntry(xy, 'f');
@@ -49,7 +47,16 @@ test('status v2: ordinary, untracked, unmerged and a stray rename entry parse wi
 		['new file.txt', 0, 2, 0],
 		['renamed.php', 1, 2, 2],
 		['conflicted.php', 1, 2, 3],
-		['gone.php', 1, 2, 0]
+		['gone.php', 1, 0, 0]
+	]);
+});
+
+test('status v2: a path removed from the index but kept on disk is one row, not two', () => {
+	// `git rm --cached keep.php` makes Git report the path twice.
+	const buf = z('1 D. N... 100644 000000 000000 hhhh 0000 keep.php', '? keep.php', '? other.txt');
+	assert.deepEqual(read.parseStatusV2Z(buf), [
+		['keep.php', 1, 2, 0],
+		['other.txt', 0, 2, 0]
 	]);
 });
 
@@ -86,6 +93,13 @@ test('cat-file --batch: hits, a missing object and a blob with newlines and NUL 
 	assert.deepEqual(answers.get('abc:empty'), Buffer.alloc(0));
 });
 
+test('cat-file --batch: a missing echo with a newline in the path stops the parse instead of poisoning it', () => {
+	const buf = Buffer.from('HEAD:b\nnope.txt missing\n1111111111111111111111111111111111111111 blob 2\nhi\n');
+	const answers = read.parseCatFileBatch(buf, ['HEAD:b\nnope.txt', 'HEAD:ok.txt']);
+	assert.equal(answers.get('HEAD:b\nnope.txt'), undefined);
+	assert.equal(answers.get('HEAD:ok.txt'), undefined);
+});
+
 test('cat-file --batch-check: oid per request, null when missing', () => {
 	const buf = Buffer.from('1111111111111111111111111111111111111111 blob 12\nabc:nope missing\n');
 	const answers = read.parseCatFileBatchCheck(buf, ['abc:package-lock.json', 'abc:nope']);
@@ -110,101 +124,21 @@ test('crlfArgs: Windows adds autocrlf only when the repository does not say', as
 	assert.equal(seen.length, 2, 'no config read off Windows');
 });
 
-// --- against a real repository ------------------------------------------
-
-function makeRepo(t) {
-	const dir = tempDir(t, 'toolkit-git-read-');
-	assert.equal(git(['init', '-b', 'trunk'], dir).status, 0);
-	fs.mkdirSync(path.join(dir, 'src'));
-	fs.writeFileSync(path.join(dir, 'src', 'wp-login.php'), '<?php // login\n');
-	fs.writeFileSync(path.join(dir, 'with space.txt'), 'spaced\n');
-	fs.writeFileSync(path.join(dir, 'package-lock.json'), '{"name":"x"}\n');
-	fs.writeFileSync(path.join(dir, '.gitignore'), 'node_modules/\n');
-	assert.equal(git(['add', '.'], dir).status, 0);
-	const commit = git(['-c', 'user.name=T', '-c', 'user.email=t@example.com', 'commit', '-q', '-m', 'first'], dir);
-	assert.equal(commit.status, 0, commit.stderr);
-	return dir;
-}
-
-test('resolveRef, readCommitInfo and currentBranch on a fresh repository', async (t) => {
-	const dir = makeRepo(t);
-	const head = git(['rev-parse', 'HEAD'], dir).stdout;
-	assert.equal(await read.resolveRef(dir, 'HEAD'), head);
-	assert.equal(await read.resolveRef(dir, 'refs/heads/trunk'), head);
-	assert.equal(await read.resolveRef(dir, 'refs/heads/nope'), null);
-
-	const info = await read.readCommitInfo(dir, 'refs/heads/trunk');
-	assert.equal(info.oid, head);
-	const seconds = Number(git(['log', '-1', '--format=%ct'], dir).stdout);
-	assert.equal(info.date, new Date(seconds * 1000).toISOString());
-	await assert.rejects(read.readCommitInfo(dir, 'refs/heads/nope'), (error) => error.code === 128);
-
-	assert.equal(await read.currentBranch(dir), 'trunk');
-	assert.equal(git(['checkout', '-q', '--detach'], dir).status, 0);
-	assert.equal(await read.currentBranch(dir), null);
+test('on Windows the autocrlf view reaches the status and diff commands themselves', async () => {
+	// A dropped spread here would be green on every other platform and bring
+	// back the ~5,000 phantom modifications of a CRLF checkout.
+	const argv = [];
+	const run = async (args) => {
+		argv.push(args);
+		return { status: args[0] === 'config' ? 1 : 0, stdout: Buffer.alloc(0), stderr: '' };
+	};
+	await read.statusRows('/site', { platform: 'win32', run });
+	await read.changesAgainst('/site', 'abc', { platform: 'win32', run });
+	const commands = argv.filter((args) => args[0] !== 'config');
+	assert.equal(commands.length, 3);
+	for (const args of commands) {
+		assert.deepEqual(args.slice(0, 2), ['-c', 'core.autocrlf=true'], args.join(' '));
+	}
+	assert.deepEqual(commands.map((args) => args[2]), ['status', 'diff', 'ls-files']);
 });
 
-test('listBranches sees a branch made by hand', async (t) => {
-	const dir = makeRepo(t);
-	assert.equal(git(['branch', 'ticket/60001'], dir).status, 0);
-	assert.deepEqual((await read.listBranches(dir)).sort(), ['ticket/60001', 'trunk']);
-});
-
-test('statusRows: clean tree, then a modification, an untracked file, a deletion, and nothing ignored', async (t) => {
-	const dir = makeRepo(t);
-	assert.deepEqual(await read.statusRows(dir, { platform: 'darwin' }), []);
-
-	fs.writeFileSync(path.join(dir, 'src', 'wp-login.php'), '<?php // edited\n');
-	fs.writeFileSync(path.join(dir, 'new file.txt'), 'new\n');
-	fs.unlinkSync(path.join(dir, 'with space.txt'));
-	fs.mkdirSync(path.join(dir, 'node_modules', 'react'), { recursive: true });
-	fs.writeFileSync(path.join(dir, 'node_modules', 'react', 'index.js'), 'expensive\n');
-
-	const rows = (await read.statusRows(dir, { platform: 'darwin' })).sort();
-	assert.deepEqual(rows, [
-		['new file.txt', 0, 2, 0],
-		['src/wp-login.php', 1, 2, 1],
-		['with space.txt', 1, 0, 1]
-	]);
-});
-
-test('changesAgainst compares the worktree with a commit that is not HEAD', async (t) => {
-	const dir = makeRepo(t);
-	const base = git(['rev-parse', 'HEAD'], dir).stdout;
-	// A second commit moves HEAD; the worktree then diverges from both.
-	fs.writeFileSync(path.join(dir, 'src', 'wp-login.php'), '<?php // parked\n');
-	assert.equal(git(['-c', 'user.name=T', '-c', 'user.email=t@example.com', 'commit', '-q', '-am', 'wip'], dir).status, 0);
-	fs.writeFileSync(path.join(dir, 'untracked.txt'), 'u\n');
-	fs.unlinkSync(path.join(dir, 'package-lock.json'));
-
-	assert.deepEqual((await read.changesAgainst(dir, 'HEAD', { platform: 'darwin' })).sort(), [
-		['package-lock.json', 1, 0, 0],
-		['untracked.txt', 0, 2, 0]
-	]);
-	assert.deepEqual((await read.changesAgainst(dir, base, { platform: 'darwin' })).sort(), [
-		['package-lock.json', 1, 0, 0],
-		['src/wp-login.php', 1, 2, 0],
-		['untracked.txt', 0, 2, 0]
-	]);
-});
-
-test('readBlobs, blobOid and treeEntryMode read one commit without touching the worktree', async (t) => {
-	const dir = makeRepo(t);
-	const base = git(['rev-parse', 'HEAD'], dir).stdout;
-	fs.writeFileSync(path.join(dir, 'src', 'wp-login.php'), '<?php // edited\n');
-
-	const blobs = await read.readBlobs(dir, base, ['src/wp-login.php', 'with space.txt', 'nope.txt']);
-	assert.equal(blobs.get('src/wp-login.php').toString('utf8'), '<?php // login\n');
-	assert.equal(blobs.get('with space.txt').toString('utf8'), 'spaced\n');
-	assert.equal(blobs.get('nope.txt'), null);
-	assert.deepEqual(await read.readBlobs(dir, base, []), new Map());
-
-	const oid = await read.blobOid(dir, base, 'package-lock.json');
-	assert.equal(oid, git(['rev-parse', `${base}:package-lock.json`], dir).stdout);
-	assert.equal(await read.blobOid(dir, base, 'nope.json'), null);
-
-	assert.equal(await read.treeEntryMode(dir, base, 'src/wp-login.php'), '100644');
-	assert.equal(await read.treeEntryMode(dir, base, 'src/nope.php'), null);
-	// The worktree edit above is still there: reads write nothing.
-	assert.equal(fs.readFileSync(path.join(dir, 'src', 'wp-login.php'), 'utf8'), '<?php // edited\n');
-});
